requirements.txt
tensorflow>=2.0.0
keras>=2.4.3
numpy>=1.19.0
pillow>=7.1.2
scikit-learn>=0.24.0
python-dotenv==0.21.0